ref!: move `copilot_cmp` to `cmp_copilot` - standard structure
